+++ app/creator/[creatorId]/page.tsx
@@ -1,0 +1,14 @@
+import SongVotingPlatform from "@/app/components/SongVotingPlatform";
+
+export default async function Page({
+    params,
+}: {
+    params: { creatorId: string };
+}) {
+    const { creatorId } = await params; 
+    return (
+        <div>
+            <SongVotingPlatform creatorId={creatorId} />
+        </div>
+    );
+}
